backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18

# AI
anthropic==0.42.0

# Statistics
scipy==1.15.1
statsmodels==0.14.4
pandas==2.2.3
numpy==2.2.2

# Charts
plotly==6.0.0
kaleido==0.2.1

# Task Queue
celery[redis]==5.4.0

# Storage (S3-compatible — optional, only needed if storage_backend=s3)
boto3==1.36.0

# Reports
weasyprint==63.1

# Validation
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
